refactor(*): rename async-hooks.js to patch-async.js
This is done in preparation of upcoming async_hooks support to make the
diff cleaner.
