+++ lib/instrumentation/index.js
@@ -40,7 +40,7 @@ Instrumentation.prototype.start = function () {
     })
   })
 
-  require('./async-hooks')(this)
+  require('./patch-async')(this)
 
   debug('shimming Module._load function')
   hook(MODULES, function (exports, name, basedir) {
