+++ lib/instrumentation/patch-async.js
